Move SingleActivity to Wizard
