+++ frog/imports/client/Wizard/components/steps/ConfigureTemplate.js
