+++ frog/imports/client/Wizard/components/steps/configureTemplate.stories.js
