+++ frog/imports/client/Wizard/containers/ConfigureTemplateContainer.js
