+++ frog/imports/client/Wizard/containers/SelectTemplateContainer.js
