+++ frog/imports/client/Wizard/store/templates.js
